fix: 다이어리 삭제기능
삭제 시 매인페이지 새로고침 및 뒤로가기 버튼 오류 수정

+++ lib/src/page/diary_page.dart
@@ -116,11 +116,11 @@ class _DiaryPageState extends State<DiaryPage> {
                       },
                     );
 
-                    if (confirmDelete) {
+                    if (confirmDelete != null && confirmDelete) {
                       await deleteDiary(widget.accessToken, widget.diaryId, () {
-                        Navigator.pop(context, true);
+                        print('Delete clicked: ${widget.diaryId}');
+                        Navigator.pop(context, true); // 삭제 성공 시 결과를 전달합니다.
                       });
-                      print('Delete clicked: ${widget.diaryId}');
                     }
                   },
                 ),
